+++ tmf630-toolkit-jpa-correlated-sort/src/test/java/org/opentmf/query/tmf630/jpa/it/SortableParentRepository.java
@@ -1,0 +1,5 @@
+package org.opentmf.query.tmf630.jpa.it;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface SortableParentRepository extends JpaRepository<SortableParent, Long> {}